refactor(app-publish/downcloud): Extract helpers
Helpers functions are now easier to mock
